+++ src/com/nentonde/praktikum/Weapon.java
@@ -4,12 +4,12 @@ public class Weapon {
 
     String weaponName;
     double weaponCombatDamage;
-    double weaponCritChance;
+    int weaponCritChance;
     double weaponMagicDamage;
     int weaponMagicUsage;
 
     //constructor
-    public Weapon(String weaponName, double weaponCombatDamage, double weaponCritChance) {
+    public Weapon(String weaponName, double weaponCombatDamage, int weaponCritChance) {
         this.weaponName = weaponName;
         this.weaponCombatDamage = weaponCombatDamage;
         this.weaponCritChance = weaponCritChance;
@@ -32,11 +32,11 @@ public class Weapon {
         this.weaponCombatDamage = weaponCombatDamage;
     }
 
-    public double getWeaponCritChance() {
+    public int getWeaponCritChance() {
         return weaponCritChance;
     }
 
-    public void setWeaponCritChance(double weaponCritChance) {
+    public void setWeaponCritChance(int weaponCritChance) {
         this.weaponCritChance = weaponCritChance;
     }
